LayoutTests/js/function-constructor-single-line-comment-expected.txt

This test checks that the Function constructor works correctly in the presence of single line comments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS (new Function('return true//'))() is true
PASS (new Function('return true;//'))() is true
PASS (new Function('a', 'return a//'))(true) is true
PASS (new Function('a', 'return a;//'))(true) is true
PASS successfullyParsed is true

TEST COMPLETE
